Fix WinGet installation issues and update pre-requisites
The Install-WinGet.ps1 script was updated to fix issues with WinGet installation. The changes include removing unnecessary code related to checking if WinGet is already installed and replacing it with a forceful installation of the required pre-requisites. This ensures that the installation process proceeds smoothly. Additionally, the script now uses the latest versions of Microsoft.VCLibs and Microsoft.UI.Xaml. These updates improve compatibility and performance. Finally, the WinGet CLI and Source Cache Package are also installed with the necessary updates.